lib/ftl: Relocation on many threads
This patch adds ability to perform band relocation
from many threads. FTL single core performance is
about one million IOPS. For random workload with
write amplification equal to 3 we need 2 million
IOPS from relocation to provide 1 million IOPS
to user.

During reloc initialization reloc pollers are started
on user defined threads. Each reloc poller is responsiable
for reading and writing relocation data. Core thread is
making decision which data should be moved and assign
this work for each reloc task.

+++ lib/ftl/ftl_band.h
@@ -173,7 +173,7 @@ struct ftl_band {
 	/* Number of open bands containing data moved from this band */
 	size_t					num_reloc_bands;
 	/* Number of blocks currently being moved from this band */
-	size_t					num_reloc_blocks;
+	int					num_reloc_blocks;
 
 	/* Free/shut bands' lists */
 	LIST_ENTRY(ftl_band)			list_entry;
